+++ Tests/RichDomain/RichDomain.Application/People/DeletePerson/DeletePersonCommandValidator.cs
@@ -1,0 +1,22 @@
+using FluentValidation;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.FluentValidation.CommandValidator", Version = "2.0")]
+
+namespace RichDomain.Application.People.DeletePerson
+{
+    public class DeletePersonCommandValidator : AbstractValidator<DeletePersonCommand>
+    {
+        [IntentManaged(Mode.Fully, Body = Mode.Merge, Signature = Mode.Merge)]
+        public DeletePersonCommandValidator()
+        {
+            ConfigureValidationRules();
+        }
+
+        [IntentManaged(Mode.Fully)]
+        private void ConfigureValidationRules()
+        {
+        }
+    }
+}