+++ Tests/RichDomain/RichDomain.Application/People/GetPeople/GetPeopleQueryValidator.cs
@@ -1,0 +1,22 @@
+using FluentValidation;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.FluentValidation.QueryValidator", Version = "2.0")]
+
+namespace RichDomain.Application.People.GetPeople
+{
+    public class GetPeopleQueryValidator : AbstractValidator<GetPeopleQuery>
+    {
+        [IntentManaged(Mode.Fully, Body = Mode.Merge, Signature = Mode.Merge)]
+        public GetPeopleQueryValidator()
+        {
+            ConfigureValidationRules();
+        }
+
+        [IntentManaged(Mode.Fully)]
+        private void ConfigureValidationRules()
+        {
+        }
+    }
+}