+++ Tests/RichDomain/RichDomain.Application/People/UpdatePersonPerson/UpdatePersonPersonCommandValidator.cs
@@ -1,0 +1,24 @@
+using FluentValidation;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.FluentValidation.CommandValidator", Version = "2.0")]
+
+namespace RichDomain.Application.People.UpdatePersonPerson
+{
+    public class UpdatePersonPersonCommandValidator : AbstractValidator<UpdatePersonPersonCommand>
+    {
+        [IntentManaged(Mode.Fully, Body = Mode.Merge, Signature = Mode.Merge)]
+        public UpdatePersonPersonCommandValidator()
+        {
+            ConfigureValidationRules();
+        }
+
+        [IntentManaged(Mode.Fully)]
+        private void ConfigureValidationRules()
+        {
+            RuleFor(v => v.FirstName)
+                .NotNull();
+        }
+    }
+}